testbed/evaluation/requirements.txt
grpcio
grpcio-tools
pyyaml
numpy
matplotlib
pandas
librosa
torch==1.13.0
torchvision==0.14